feat: 포스트 조회 기능 구현완료

+++ backend/src/main/java/com/kkini/core/domain/comment/entity/Comment.java
@@ -25,9 +25,7 @@ public class Comment extends BaseEntityWithCreatedTime {
     @JoinColumn(name = "post_id")
     private Post post;
 
-//    @ManyToOne
-//    @JoinColumn(name = "parents_id")
-    private int parentsId;
+    private Long parentsId;
 
     private String contents;
